Updated reflection and json
