+++ Core/DataStructure/NestedString.cs
@@ -3,8 +3,6 @@ using System.Text;
 
 namespace Cutulu.Core
 {
-    using Reflection;
-
     public class NestedString
     {
         public List<NestedString> Children { get; set; }
